requirements.txt
streamlit
numpy
matplotlib
gTTS
soundfile
openai-whisper
torch